Eingaben sind nun möglich

+++ classes/helper/Dbhelper.php
@@ -13,19 +13,25 @@ abstract class Dbhelper{
 	protected $id;	//db id
 	protected $dbh=null; //databasehandle
 
-	protected $tablename; 
+	protected $tablename=''; 
 	
-	public function  superDelete(){
+	public function  mysqlDelete(){		
+		
+		if(!$this->dbh() || !is_numeric($this->id()) || $this->tablename=='')
+			return false;
+		echo "<h3>in</h3>";
 		$sql="DELETE FROM {$this->tablename} WHERE id=?";
+		$stmt=$this->dbh->prepare($sql);
+		
+		$stmt->bindParam(1, $this->id, PDO::PARAM_INT);
+		$stmt->execute();		
 		return $sql;
 	}
 	protected function setTableName($name){
 		$this->tablename=$name;
 		
-	}
-	/**SetGetter**/
-
-	public function id($id=""){ if (empty($id)){return $this->id; }else{$this->id=$id; } }
+	}		
+	
 	public function dbh(PDO $dbh=null){
 		if (empty($dbh)){
 			if($this->dbh === null)
@@ -35,5 +41,9 @@ abstract class Dbhelper{
 			$this->dbh=$dbh;
 		}
 	}//dbh
+	
+	/**SetGetter**/
+	public function id($id=""){ if (empty($id)){return $this->id; }else{$this->id=$id; } }
+	
 
 }//class
